+++ 프로그래밍 응용 과제/Chapter 5/Chapter 5 - Library/Date.h
@@ -10,7 +10,7 @@ private:
 public:
 	Date(std::string dateStr);
 	
-	int totalDays();
+	long totalDays();
 	std::string to_string();
 };
